Updated to support A-Frame 0.7 (#29)
* Checking canvasEl exists before removing eventlisteners

In certain cases, especially when using React, the canvasEl may have been removed before removeEventListeners gets called. I’ve added a check for this so errors aren’t thrown if this is the case.

* Updated dist

* Updated to support A-Frame 0.7

Switched from using `render-target-loaded` to `renderstart` and updated dist. This correctly supports the breaking changes to the way canvas is initialized in A-Frame 0.7 and later versions of THREE.js, allowing event listeners to be bound at the correct time.